requirements.txt
# Core dependencies for MCP playground
requests>=2.31.0
yfinance>=0.2.0
python-dotenv>=1.0.0

